+++ lib/src/data/pagination_data_wrapper.dart
@@ -1,8 +1,0 @@
-import 'package:april_flutter_utils/src/data/refreshable_data_wrapper.dart';
-
-///分页数据结构包装
-abstract class AbsPaginationDataWrapper<T>
-    extends AbsRefreshableDataWrapper<T> {
-  ///是否还有更多数据
-  bool get hasMore;
-}