+++ lib/src/data/refreshable/refreshable_controller.dart
@@ -1,7 +1,4 @@
-import 'package:flutter/foundation.dart';
-
-import 'package:april_flutter_utils/src/data/refreshable.dart';
-import 'refreshable_data_wrapper.dart';
+part of 'refreshable.dart';
 
 ///刷新列表控制器
 abstract class AbsRefreshableController<T, W extends AbsRefreshableDataWrapper<T>> extends Refreshable<T> {
@@ -42,11 +39,11 @@ abstract class AbsRefreshableController<T, W extends AbsRefreshableDataWrapper<T
     if (refreshableStateInternal.value.isRefreshing) {
       return;
     }
-    await _paginationRefresh();
+    await _refresh();
   }
 
   ///刷新操作真实执行函数
-  Future<void> _paginationRefresh() async {
+  Future<void> _refresh() async {
     setRefreshing(true);
     //加载数据
     try {
@@ -108,28 +105,4 @@ abstract class AbsRefreshableController<T, W extends AbsRefreshableDataWrapper<T
   ///刷新操作的具体执行方法
   @protected
   Future<W> refreshInternal();
-}
-
-///刷新功能的一些配置参数
-class RefreshableConfig {
-  const RefreshableConfig({
-    this.notifyStateFirst = true,
-    this.autoRefresh = true,
-    this.lazyRefresh = true,
-    this.autoRefreshOnEmptyList = true,
-  });
-
-  ///是否优先更新状态，而不是数据
-  final bool notifyStateFirst;
-
-  ///是否自动触发一次刷新
-  final bool autoRefresh;
-
-  ///是否懒刷新，如果是，则在第一次获取数据监听器的时候刷新
-  ///否则，则在构造函数内刷新。
-  ///当 [autoRefresh] 为 true 时生效
-  final bool lazyRefresh;
-
-  ///获取数据监听器其时，发现数据量为空，是否触发刷新
-  final bool autoRefreshOnEmptyList;
 }